Update iredmail.sh
Added check for $NTPSERVER -- if exist, comment out the 4 default pool servers and add the one server. (In my use case, there is a single NTP server in a vPod environment, and ALL TIME in the pod should point only there since there is no guarantee that the pod will be connected to the internet once it is put to use.